Refactoring of component names and structure
