+++ client/index.html
@@ -42,7 +42,7 @@
     <!-- build:js(client) scripts/lt-ie9.js -->
     <!--[if lt IE 9]>
     <script src="/bower_components/es5-shim/es5-shim.js"></script>
-    <script src="/bower_components/es5-shim/es5-sham.js"></script>
+
     <![endif]-->
     <!-- endbuild -->
 
@@ -56,11 +56,10 @@
       <script src="bower_components/underscore/underscore.js"></script>
       <script src="bower_components/i18next/i18next.js"></script>
       <script src="bower_components/scooch/build/scooch.min.js"></script>
-      <script src="bower_components/esri-leaflet-renderers/dist/esri-leaflet-renderers.js"></script>
-
-    <!-- endbower -->
+      <!-- endbower -->
     <!-- endbuild -->
 
+    <script src="bower_components/esri-leaflet-renderers/dist/esri-leaflet-renderers.js"></script>
     <script src="scripts/main.js"></script>
   </body>
 </html>
