+++ client/scripts/components/footer.jsx
@@ -2,7 +2,7 @@
 
 var React = require('react')
 var RouteHandler = require('react-router').RouteHandler;
-var Message=require('./lanMessage.jsx');
+
 module.exports = React.createClass({
   render: function() {
     return (
