+++ client/scripts/main.js
@@ -7,7 +7,8 @@ window.ESRI_SELF_URL='https://www.arcgis.com/sharing/rest/portals/self';
 
 window.ESRI_PROXY_URL = 'http://localhost:3553'
 window.ESIR_USE_PROXY = true;
-
+//var Message=require('./components/commons/message.jsx');	
+window.Message=require('./components/commons/message.jsx');
 var React = require('react');
 var Router = require('react-router');
 var router = require('./router.jsx');
@@ -21,6 +22,7 @@ var options = {
 
 window.i18n.init(options, function (t) {
 	router.run(function (Handler, state) {
+
 		React.render(React.createElement(Handler, state), document.getElementById('app-wrapper'));
 	});
 });
