+++ client/scripts/stores/arcgisLayerStore.js
