+++ client/scripts/stores/arcgisLoginStore.js
