+++ client/styles/main.scss
@@ -85,3 +85,4 @@
       }
     }
   }
+}
